+++ changelog.md
@@ -27,6 +27,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Changed MSG_TYPE_38 to MSG_TYPE_REGISTER_STATUS
 - Change heater on/off to On/Off to match other values in /status page
+- Improved the internal locking/mutex for MQTT publishing
 
 ### Fixed
 - Publish ORP and PH setpoint to MQTT